Comment "Fixes bug: start always evaluates to true"
Note: that after this commit, closing the LMS driver makes the test program hang.
(it blocks on the Read). At the end of the test program, start == 0 as argument.
so the branch is always executed anyway.. Maybe because it didn't started with it?
I'm also unsure what we are trying to do here.. empty the socket buffer? (why care anyway?)